# backend/endpoints/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SaveForge 2.0 GameCatalog getters",
    "version": "1.0.0",
    "description": "Local, read-only HTTP surface over the currently implemented GameCatalog getters. Every operation calls exactly one Go getter and serialises its typed result. Getters whose contract file declares 'contract definition only' (GetResources) are not exposed."
  },
  "servers": [
    {
      "url": "/",
      "description": "Local explorer"
    }
  ],
  "paths": {
    "/api/v1/catalog/info": {
      "get": {
        "operationId": "getCatalogInfo",
        "summary": "GetCatalogInfo",
        "description": "Returns the manifest of the loaded GameCatalog: schema version, data version, game version, validation status and the used sources.",
        "responses": {
          "200": {
            "description": "Catalog manifest",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCatalogInfoResult"
                }
              }
            }
          },
          "500": {
            "description": "The loaded catalog manifest is missing or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalog/resource": {
      "get": {
        "operationId": "getResource",
        "summary": "GetResource",
        "description": "Returns one full catalog resource document. The resource identity is the exact pair (kind, key); the kind is resolved first and the key only inside that kind.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ResourceKind"
          },
          {
            "$ref": "#/components/parameters/ResourceKey"
          }
        ],
        "responses": {
          "200": {
            "description": "The resource document",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetResourceResult"
                }
              }
            }
          },
          "400": {
            "description": "kind or key is missing, or the pair is unknown. The message is the one the getter returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalog/resource-relations": {
      "get": {
        "operationId": "getResourceRelations",
        "summary": "GetResourceRelations",
        "description": "Returns the outgoing and incoming relations of one resource identified by the exact pair (kind, key), in catalog order. Relations only: the documents of the related resources belong to GetResource. A filtered-out direction stays an empty array, never null.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ResourceKind"
          },
          {
            "$ref": "#/components/parameters/ResourceKey"
          },
          {
            "$ref": "#/components/parameters/RelationType"
          },
          {
            "$ref": "#/components/parameters/RelationDirection"
          }
        ],
        "responses": {
          "200": {
            "description": "The relations of the resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetResourceRelationsResult"
                }
              }
            }
          },
          "400": {
            "description": "kind or key is missing, the pair is unknown, or relationType or direction is not supported. The message is the one the getter returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalog/item-variants": {
      "get": {
        "operationId": "getItemVariants",
        "summary": "GetItemVariants",
        "description": "Returns the variants stored in one item document, in catalog order. An item without variants returns an empty list; the base item is never synthesised into a variant.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ResourceKind"
          },
          {
            "$ref": "#/components/parameters/ResourceKey"
          }
        ],
        "responses": {
          "200": {
            "description": "The item variants",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetItemVariantsResult"
                }
              }
            }
          },
          "400": {
            "description": "kind or key is missing, kind is not item, or the pair is unknown. The message is the one the getter returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/healthz": {
      "get": {
        "operationId": "healthz",
        "summary": "Liveness probe",
        "responses": {
          "200": {
            "description": "The explorer is running",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    }
                  },
                  "required": ["status"]
                }
              }
            }
          }
        }
      }
    },
    "/openapi.json": {
      "get": {
        "operationId": "openapiDocument",
        "summary": "This OpenAPI document",
        "responses": {
          "200": {
            "description": "OpenAPI 3.0 specification",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/docs": {
      "get": {
        "operationId": "docsUI",
        "summary": "Local browser explorer",
        "responses": {
          "200": {
            "description": "HTML explorer page",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "ResourceKind": {
        "name": "kind",
        "in": "query",
        "required": true,
        "description": "schema.Resource.Kind, matched exactly. The kind is resolved before the key.",
        "schema": {
          "type": "string"
        },
        "example": "item"
      },
      "ResourceKey": {
        "name": "key",
        "in": "query",
        "required": true,
        "description": "schema.Resource.Key inside the given kind, matched exactly. An item key is eight uppercase hexadecimal characters without a kind prefix. Not an Item.GameID, not a variant ID and not a data file name.",
        "schema": {
          "type": "string"
        },
        "example": "000F4240"
      },
      "RelationType": {
        "name": "relationType",
        "in": "query",
        "required": false,
        "description": "schema.RelationKind, matched exactly and case-sensitively. Omitted or empty means every relation type.",
        "schema": {
          "type": "string",
          "enum": ["", "compatible_with_aow", "requires_container"]
        },
        "example": "compatible_with_aow"
      },
      "RelationDirection": {
        "name": "direction",
        "in": "query",
        "required": false,
        "description": "Relation direction, matched exactly. Omitted or empty means both directions; there is deliberately no 'both' value. The filtered-out direction stays an empty array.",
        "schema": {
          "type": "string",
          "enum": ["", "outgoing", "incoming"]
        },
        "example": "outgoing"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "description": "Error envelope carrying the message returned by the getter.",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": ["error"]
      },
      "DataSource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "evidence": {
            "type": "string"
          },
          "reviewed": {
            "type": "boolean"
          }
        },
        "required": ["id", "kind", "location", "version", "evidence", "reviewed"]
      },
      "GetCatalogInfoResult": {
        "type": "object",
        "properties": {
          "schemaVersion": {
            "type": "integer",
            "format": "int64"
          },
          "dataVersion": {
            "type": "string"
          },
          "gameVersion": {
            "type": "string"
          },
          "valid": {
            "type": "boolean",
            "description": "True only means schema.ValidateManifest accepted the manifest."
          },
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataSource"
            }
          }
        },
        "required": ["schemaVersion", "dataVersion", "gameVersion", "valid", "sources"]
      },
      "Resource": {
        "type": "object",
        "description": "schema.Resource identified by the (kind, key) pair. The item document is passed through as stored in the catalog; its full shape is owned by the catalog schema, not by this explorer.",
        "properties": {
          "key": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "item": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true
          }
        },
        "required": ["key", "kind", "item"]
      },
      "ItemVariant": {
        "type": "object",
        "description": "schema.ItemVariant as stored in the catalog, with full provenance.",
        "additionalProperties": true
      },
      "GetResourceResult": {
        "type": "object",
        "properties": {
          "resource": {
            "$ref": "#/components/schemas/Resource"
          }
        },
        "required": ["resource"]
      },
      "ResourceRef": {
        "type": "object",
        "description": "schema.ResourceRef: the (kind, key) pair that identifies a resource. There is no numeric resource identifier.",
        "properties": {
          "kind": {
            "type": "string",
            "example": "item"
          },
          "key": {
            "type": "string",
            "example": "000F4240"
          }
        },
        "required": ["kind", "key"]
      },
      "Relation": {
        "type": "object",
        "description": "schema.Relation as derived and stored by the catalog. Both endpoints are ResourceRef objects, never numeric IDs.",
        "properties": {
          "from": {
            "$ref": "#/components/schemas/ResourceRef"
          },
          "to": {
            "$ref": "#/components/schemas/ResourceRef"
          },
          "kind": {
            "type": "string",
            "enum": ["compatible_with_aow", "requires_container"]
          },
          "provenance": {
            "type": "object",
            "description": "The evidence the catalog derived the relation from.",
            "additionalProperties": true
          }
        },
        "required": ["from", "to", "kind", "provenance"]
      },
      "GetResourceRelationsResult": {
        "type": "object",
        "properties": {
          "outgoing": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Relation"
            }
          },
          "incoming": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Relation"
            }
          }
        },
        "required": ["outgoing", "incoming"]
      },
      "GetItemVariantsResult": {
        "type": "object",
        "properties": {
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemVariant"
            }
          }
        },
        "required": ["variants"]
      }
    }
  }
}